refactor: Renamed file, updated styling
