+++ text2sql-frontend/src/hooks/index.ts
@@ -1,4 +1,4 @@
-export * from "./auth";
+export * from "./settings";
 export * from "./connections";
 export * from "./conversations";
 export * from "./tableSchema";
