+++ text2sql-frontend/src/hooks/messages.ts
@@ -8,7 +8,7 @@ import {
   useQueryClient,
 } from "@tanstack/react-query";
 import { enqueueSnackbar } from "notistack";
-import { getBackendStatusQuery } from "@/hooks/auth";
+import { getBackendStatusQuery } from "@/hooks/settings";
 
 const MESSAGES_QUERY_KEY = ["MESSAGES"];
 const QUERIES_QUERY_KEY = ["SQL_QUERIES"];
